# openspec/openapi-v1.yaml
openapi: 3.1.0
info:
  title: AstrBot OpenAPI v1
  version: 0.1.0
  description: >
    Target REST contract for migrating AstrBot HTTP APIs to /api/v1.
    Dynamic AstrBot configuration payloads are intentionally modeled as open
    JSON objects because their schemas are provided at runtime by template
    endpoints.
    Developer API keys currently support these scopes only: bot, provider,
    persona, im, config, chat, data, file, plugin, mcp, skill, plus the sensitive
    config:edit_admin and chat:admin sub-scopes. The config scope also grants bot
    and provider access, but no parent scope grants a sensitive sub-scope.
...
servers:
  - url: http://localhost:6185
    description: Local AstrBot server
security:
  - ApiKeyAuth: []
...
paths:
...
  /api/v1/chat/projects:
    get:
      tags: [Chat]
      summary: List ChatUI projects
      operationId: listChatProjects
      x-astrbot-scope: chat
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    post:
      tags: [Chat]
      summary: Create a ChatUI project
      operationId: createChatProject
      x-astrbot-scope: chat
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatProjectRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/projects/{project_id}:
    get:
      tags: [Chat]
      summary: Get a ChatUI project
      operationId: getChatProject
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    patch:
      tags: [Chat]
      summary: Update a ChatUI project
      operationId: updateChatProject
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChatProjectRequest"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
    delete:
      tags: [Chat]
      summary: Delete a ChatUI project
      operationId: deleteChatProject
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/projects/{project_id}/sessions:
    get:
      tags: [Chat]
      summary: List sessions in a ChatUI project
      operationId: listChatProjectSessions
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/projects/{project_id}/workspace/files:
    get:
      tags: [Chat]
      summary: List files in a ChatUI project workspace directory
      operationId: listChatProjectWorkspaceFiles
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: false
          schema:
            type: string
            default: ""
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/projects/{project_id}/workspace/file:
    get:
      tags: [Chat]
      summary: Read a file in a ChatUI project workspace
      operationId: getChatProjectWorkspaceFile
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/projects/{project_id}/workspace/file/download:
    get:
      tags: [Chat]
      summary: Download a file from a ChatUI project workspace
      operationId: downloadChatProjectWorkspaceFile
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - name: path
          in: query
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Workspace file content
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
  /api/v1/chat/projects/{project_id}/sessions/{session_id}:
    post:
      tags: [Chat]
      summary: Add a session to a ChatUI project
      operationId: addChatProjectSession
      x-astrbot-scope: chat
      parameters:
        - name: project_id
          in: path
          required: true
          schema:
            type: string
        - $ref: "#/components/parameters/SessionId"
        - name: position
          in: query
          required: false
          description: 0-based insertion index in the project session list (omitted prepends to top)
          schema:
            type: integer
      responses:
        "200":
          $ref: "#/components/responses/Ok"
  /api/v1/chat/projects/sessions/{session_id}:
    delete:
      tags: [Chat]
      summary: Remove a session from its ChatUI project
      operationId: removeChatProjectSession
      x-astrbot-scope: chat
      parameters:
        - $ref: "#/components/parameters/SessionId"
      responses:
        "200":
          $ref: "#/components/responses/Ok"
...
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    Ok:
      description: Standard AstrBot success response
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/SuccessEnvelope"
...
  parameters:
...
    SessionId:
      name: session_id
      in: path
      required: true
      schema:
        type: string
...
  schemas:
    SuccessEnvelope:
      type: object
      required: [status, data]
      properties:
        status:
          type: string
          const: ok
        message:
          type: string
        data: {}
      additionalProperties: false
...
    ChatProjectRequest:
      type: object
      properties:
        title:
          type: string
        emoji:
          type: string
        description:
          type: string
        workspace_type:
          type: string
          enum: [session, project, custom]
          description: Workspace mode. API key callers may use only session or project; project is the default.
        workspace_path:
          type: string
          description: Dashboard-only custom workspace path. API key callers cannot set this field.
        spcode_auto_load:
          type: boolean
          nullable: true
          default: true
        spcode_force:
          type: boolean
          nullable: true
          default: false
        spcode_no_codegraph:
          type: boolean
          nullable: true
          default: false
      additionalProperties: false
...